% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raztr_path.R
\name{get_raztr_path}
\alias{get_raztr_path}
\title{Get the full path of a file in the \code{inst/extdata} folder}
\usage{
get_raztr_path(filename)
}
\arguments{
\item{filename}{name of a file}
}
\value{
the full path of the filename, if and only if
the file is present. Will stop otherwise.
}
\description{
Get the full path of a file in the \code{inst/extdata} folder
}
\examples{
library(testthat)

expect_true(file.exists(get_raztr_path("parameters.RDa")))
expect_error(get_raztr_path("abs.ent"))
}
